chore: Change type 'ID & UserID' uint to uint64 in 'task entity' from domain layer

+++ internal/domain/entity/task.go
@@ -3,8 +3,8 @@ package entity
 import "time"
 
 type Task struct {
-	ID          uint
-	UserID      uint
+	ID          uint64
+	UserID      uint64
 	Title       string
 	Description string
 	Status      TaskStatus
